api/requirements.txt
authlib==0.11
flask==1.0.2
pymongo==3.6.1
google-api-python-client
google-auth
virtualenv